requirements.txt
# requirements.txt
PyQt5
openai
python-dotenv